delete added line
I put one line in the program to check that my ability to clone it to the dev folder, edit from my local repo and then push it to the remote repo on my command line. It worked! Now I am confirming that when I delete the line here on github that I can pull the new version down.